+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 
 - Fixed an issue where an `Announcement` object sometimes didn't have an associated course ID. (Thanks, [@bennettscience](https://github.com/bennettscience))
 
+### New Endpoint Coverage
+
+- Course TODO items (Thanks, [@onomou](https://github.com/onomou))
+- Create observer pairing code (Thanks, [@bennettscience](https://github.com/bennettscience))
+
 ## [2.0.0] - 2020-08-14
 
 ### General