+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added missing documentation for the get_current_user method and clarifications to the CurrentUser class. (Thanks, [@Xx-Ashutosh-xX](https://github.com/Xx-Ashutosh-xX))
 - `Canvas.get_announcement` now has a required parameter `context_codes`, which accepts a list of course IDs or `Course` objects.
+- Added basic troubleshooting guide to documentation
 
 ### Bugfixes